requirements.txt
tkinter
cryptography
pillow